Fix Visual C++ compiler warning
When declaring number constants using the scientific notation, they are of type double. Added a
static_cast to satisfy the Visual C++ compiler.

+++ src/LoopChecker.cpp
@@ -4,7 +4,7 @@
 
 #include "LoopChecker.hpp"
 
-const uint64_t SEC_TO_MICRO = 1e6;
+const uint64_t SEC_TO_MICRO = static_cast<uint64_t>(1e6);
 
 namespace nr {
